# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "gin-demo",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/api/v1",
    "paths": {
        "/collectors": {
            "get": {
                "description": "查询采集器列表",
                "tags": [
                    "Collector"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "PageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "PageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "创建采集器",
                "tags": [
                    "Collector"
                ],
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CollectorDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/collectors/{id}": {
            "delete": {
                "description": "删除采集器",
                "tags": [
                    "Collector"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "采集器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/projects": {
            "get": {
                "description": "查询项目列表",
                "tags": [
                    "Project"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作描述",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "页数",
                        "name": "page",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "数据条数",
                        "name": "size",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "查询用户列表",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "操作者",
                        "name": "operator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "操作描述",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CollectorDto": {
            "type": "object",
            "required": [
                "cluster",
                "envId",
                "name",
                "planId",
                "projectId"
            ],
            "properties": {
                "alertCnt": {
                    "type": "integer"
                },
                "alertUrl": {
                    "type": "string"
                },
                "appId": {
                    "type": "integer"
                },
                "cluster": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "integer"
                },
                "envId": {
                    "type": "integer"
                },
                "envName": {
                    "type": "string"
                },
                "execCnt": {
                    "type": "integer"
                },
                "heartBeatInterval": {
                    "type": "integer"
                },
                "heartBeatUrl": {
                    "type": "string"
                },
                "metricUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "planId": {
                    "type": "integer"
                },
                "projectCode": {
                    "type": "string"
                },
                "projectId": {
                    "type": "integer"
                },
                "tokenRefreshInterval": {
                    "type": "integer"
                },
                "tokenUrl": {
                    "type": "string"
                }
            }
        }
    }
}